src/main/java/dnf/itemauction/domain/login/LoginService.java
package dnf.itemauction.domain.login;

import dnf.itemauction.domain.member.Member;
import dnf.itemauction.domain.member.MemberRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class LoginService {

    private final MemberRepository memberRepository;

    public Member login(String loginId, String password){
        return memberRepository.findByLoginId(loginId)
                .filter(member -> member.getPassword().equals(password))
                .orElse(null);
    }
}
